return errors instead of duplicate xml/json
check_parameter will return errors as part of the function

+++ src/Module/Api/Subsonic_Api.php
@@ -528,7 +528,7 @@ class Subsonic_Api
      */
     public static function getmusicdirectory($input)
     {
-        $object_id = self::_check_parameter($input, 'id');
+        $object_id = $input['id'] ?? 0;
         $response  = Subsonic_Xml_Data::addSubsonicResponse('getmusicdirectory');
         if ((int)$object_id === 0) {
             $response = Subsonic_Xml_Data::addError(Subsonic_Xml_Data::SSERROR_DATA_NOTFOUND, "Directory not found", 'getmusicdirectory');
@@ -589,7 +589,9 @@ class Subsonic_Api
     public static function getartist($input)
     {
         $artistid = self::_check_parameter($input, 'id');
-
+        if (!$artistid) {
+            return;
+        }
         $artist = new Artist(Subsonic_Xml_Data::_getAmpacheId($artistid));
         if ($artist->isNew()) {
             $response = Subsonic_Xml_Data::addError(Subsonic_Xml_Data::SSERROR_DATA_NOTFOUND, "Artist not found.", 'getartist');
@@ -610,7 +612,10 @@ class Subsonic_Api
     public static function getalbum($input)
     {
         $albumid = self::_check_parameter($input, 'id');
-        $album   = new Album(Subsonic_Xml_Data::_getAmpacheId($albumid));
+        if (!$albumid) {
+            return;
+        }
+        $album = new Album(Subsonic_Xml_Data::_getAmpacheId($albumid));
         if (!isset($album->id)) {
             $response = Subsonic_Xml_Data::addError(Subsonic_Xml_Data::SSERROR_DATA_NOTFOUND, "Album not found.", 'getalbum');
         } else {
@@ -631,7 +636,10 @@ class Subsonic_Api
      */
     public static function getsong($input)
     {
-        $songid   = self::_check_parameter($input, 'id');
+        $songid = self::_check_parameter($input, 'id');
+        if (!$songid) {
+            return;
+        }
         $response = Subsonic_Xml_Data::addSubsonicResponse('getsong');
         $song     = Subsonic_Xml_Data::_getAmpacheId($songid);
         Subsonic_Xml_Data::addSong($response, $song);
@@ -674,12 +682,15 @@ class Subsonic_Api
      */
     public static function getartistinfo($input, $elementName = "artistInfo")
     {
-        $id                = self::_check_parameter($input, 'id');
+        $object_id = self::_check_parameter($input, 'id');
+        if (!$object_id) {
+            return;
+        }
         $count             = $input['count'] ?? 20;
         $includeNotPresent = (array_key_exists('includeNotPresent', $input) && $input['includeNotPresent'] === "true");
 
-        if (Subsonic_Xml_Data::_isArtist($id)) {
-            $artist_id = Subsonic_Xml_Data::_getAmpacheId($id);
+        if (Subsonic_Xml_Data::_isArtist($object_id)) {
+            $artist_id = Subsonic_Xml_Data::_getAmpacheId($object_id);
             $info      = Recommendation::get_artist_info($artist_id);
             $similars  = Recommendation::get_artists_like($artist_id, $count, !$includeNotPresent);
             $response  = Subsonic_Xml_Data::addSubsonicResponse($elementName);
@@ -749,11 +760,14 @@ class Subsonic_Api
             return;
         }
 
-        $id    = self::_check_parameter($input, 'id');
+        $object_id = self::_check_parameter($input, 'id');
+        if (!$object_id) {
+            return;
+        }
         $count = $input['count'] ?? 50;
         $songs = array();
-        if (Subsonic_Xml_Data::_isArtist($id)) {
-            $similars = Recommendation::get_artists_like(Subsonic_Xml_Data::_getAmpacheId($id));
+        if (Subsonic_Xml_Data::_isArtist($object_id)) {
+            $similars = Recommendation::get_artists_like(Subsonic_Xml_Data::_getAmpacheId($object_id));
             if (!empty($similars)) {
                 debug_event(self::class, 'Found: ' . count($similars) . ' similar artists', 4);
                 foreach ($similars as $similar) {
@@ -771,9 +785,9 @@ class Subsonic_Api
             // randomize and slice
             shuffle($songs);
             $songs = array_slice($songs, 0, $count);
-        //} elseif (Ampache\Module\Api\Subsonic_Xml_Data::_isAlbum($id)) { // TODO: support similar songs for albums
-        } elseif (Subsonic_Xml_Data::_isSong($id)) {
-            $songs = Recommendation::get_songs_like(Subsonic_Xml_Data::_getAmpacheId($id), $count);
+        //} elseif (Ampache\Module\Api\Subsonic_Xml_Data::_isAlbum($object_id)) { // TODO: support similar songs for albums
+        } elseif (Subsonic_Xml_Data::_isSong($object_id)) {
+            $songs = Recommendation::get_songs_like(Subsonic_Xml_Data::_getAmpacheId($object_id), $count);
         }
 
         if (count($songs) == 0) {
@@ -1165,18 +1179,19 @@ class Subsonic_Api
      */
     public static function getplaylist($input)
     {
-        $playlistid = self::_check_parameter($input, 'id');
-        if ($playlistid) {
-            $response = Subsonic_Xml_Data::addSubsonicResponse('getplaylist');
-            if (Subsonic_Xml_Data::_isSmartPlaylist($playlistid)) {
-                $playlist = new Search(Subsonic_Xml_Data::_getAmpacheId($playlistid), 'song');
-                Subsonic_Xml_Data::addPlaylist($response, $playlist, true);
-            } else {
-                $playlist = new Playlist(Subsonic_Xml_Data::_getAmpacheId($playlistid));
-                Subsonic_Xml_Data::addPlaylist($response, $playlist, true);
-            }
-            self::_apiOutput($input, $response);
+        $playlistId = self::_check_parameter($input, 'id');
+        if (!$playlistId) {
+            return;
         }
+        $response = Subsonic_Xml_Data::addSubsonicResponse('getplaylist');
+        if (Subsonic_Xml_Data::_isSmartPlaylist($playlistId)) {
+            $playlist = new Search(Subsonic_Xml_Data::_getAmpacheId($playlistId), 'song');
+            Subsonic_Xml_Data::addPlaylist($response, $playlist, true);
+        } else {
+            $playlist = new Playlist(Subsonic_Xml_Data::_getAmpacheId($playlistId));
+            Subsonic_Xml_Data::addPlaylist($response, $playlist, true);
+        }
+        self::_apiOutput($input, $response);
     }
 
     /**
@@ -1256,7 +1271,9 @@ class Subsonic_Api
     public static function deleteplaylist($input)
     {
         $playlistId = self::_check_parameter($input, 'id');
-
+        if (!$playlistId) {
+            return;
+        }
         if (Subsonic_Xml_Data::_isSmartPlaylist($playlistId)) {
             $playlist = new Search(Subsonic_Xml_Data::_getAmpacheId($playlistId), 'song');
         } else {
@@ -1611,9 +1628,11 @@ class Subsonic_Api
     public static function setrating($input)
     {
         $object_id = self::_check_parameter($input, 'id');
-        $rating    = $input['rating'];
-
-        $robj = null;
+        if (!$object_id) {
+            return;
+        }
+        $rating = $input['rating'];
+        $robj   = null;
         if (Subsonic_Xml_Data::_isArtist($object_id)) {
             $robj = new Rating(Subsonic_Xml_Data::_getAmpacheId($object_id), "artist");
         } else {
@@ -1646,6 +1665,9 @@ class Subsonic_Api
     public static function scrobble($input)
     {
         $object_ids = self::_check_parameter($input, 'id');
+        if (!$object_ids) {
+            return;
+        }
         $submission = (array_key_exists('submission', $input) && ($input['submission'] === 'true' || $input['submission'] === '1'));
         $username   = (string)$input['u'];
         $client     = (string)filter_var($input['c'], FILTER_SANITIZE_STRING) ?? 'Subsonic';
@@ -1712,25 +1734,28 @@ class Subsonic_Api
      */
     public static function createshare($input)
     {
-        $libitem_id  = self::_check_parameter($input, 'id');
+        $object_id = self::_check_parameter($input, 'id');
+        if (!$object_id) {
+            return;
+        }
         $description = $input['description'];
         if (AmpConfig::get('share')) {
             $expire_days = Share::get_expiry($input['expires'] ?? null);
             $object_type = null;
-            if (is_array($libitem_id) && Subsonic_Xml_Data::_isSong($libitem_id[0])) {
-                $song_id     = Subsonic_Xml_Data::_getAmpacheId($libitem_id[0]);
+            if (is_array($object_id) && Subsonic_Xml_Data::_isSong($object_id[0])) {
+                $song_id     = Subsonic_Xml_Data::_getAmpacheId($object_id[0]);
                 $tmp_song    = new Song($song_id);
                 $object_id   = Subsonic_Xml_Data::_getAmpacheId($tmp_song->album);
                 $object_type = 'album';
             } else {
-                $object_id = Subsonic_Xml_Data::_getAmpacheId($libitem_id);
-                if (Subsonic_Xml_Data::_isAlbum($libitem_id)) {
+                $object_id = Subsonic_Xml_Data::_getAmpacheId($object_id);
+                if (Subsonic_Xml_Data::_isAlbum($object_id)) {
                     $object_type = 'album';
                 }
-                if (Subsonic_Xml_Data::_isSong($libitem_id)) {
+                if (Subsonic_Xml_Data::_isSong($object_id)) {
                     $object_type = 'song';
                 }
-                if (Subsonic_Xml_Data::_isPlaylist($libitem_id)) {
+                if (Subsonic_Xml_Data::_isPlaylist($object_id)) {
                     $object_type = 'playlist';
                 }
             }
@@ -1771,8 +1796,11 @@ class Subsonic_Api
      */
     public static function updateshare($input)
     {
-        $username    = self::_check_parameter($input, 'u');
-        $share_id    = self::_check_parameter($input, 'id');
+        $username = self::_check_parameter($input, 'u');
+        $share_id = self::_check_parameter($input, 'id');
+        if (!$share_id) {
+            return;
+        }
         $user        = User::get_from_username((string)$username);
         $description = $input['description'];
 
@@ -1824,9 +1852,12 @@ class Subsonic_Api
     {
         $username = self::_check_parameter($input, 'u');
         $user     = User::get_from_username((string)$username);
-        $id       = self::_check_parameter($input, 'id');
+        $share_id = self::_check_parameter($input, 'id');
+        if (!$share_id) {
+            return;
+        }
         if (AmpConfig::get('share')) {
-            if (Share::delete_share((int) $id, $user)) {
+            if (Share::delete_share((int) $share_id, $user)) {
                 $response = Subsonic_Xml_Data::addSubsonicResponse('deleteshare');
             } else {
                 $response = Subsonic_Xml_Data::addError(Subsonic_Xml_Data::SSERROR_DATA_NOTFOUND, '', 'deleteshare');
@@ -1951,7 +1982,9 @@ class Subsonic_Api
     public static function deletepodcastchannel($input)
     {
         $podcast_id = (int)self::_check_parameter($input, 'id');
-
+        if (!$podcast_id) {
+            return;
+        }
         if (AmpConfig::get('podcast') && Access::check('interface', 75)) {
             $podcast = new Podcast(Subsonic_Xml_Data::_getAmpacheId($podcast_id));
             if ($podcast->id) {
@@ -1977,10 +2010,12 @@ class Subsonic_Api
      */
     public static function deletepodcastepisode($input)
     {
-        $id = self::_check_parameter($input, 'id');
-
+        $episode_id = self::_check_parameter($input, 'id');
+        if (!$episode_id) {
+            return;
+        }
         if (AmpConfig::get('podcast') && Access::check('interface', 75)) {
-            $episode = new Podcast_Episode(Subsonic_Xml_Data::_getAmpacheId($id));
+            $episode = new Podcast_Episode(Subsonic_Xml_Data::_getAmpacheId($episode_id));
             if ($episode->id !== null) {
                 if ($episode->remove()) {
                     $response = Subsonic_Xml_Data::addSubsonicResponse('deletepodcastepisode');
@@ -2005,10 +2040,12 @@ class Subsonic_Api
      */
     public static function downloadpodcastepisode($input)
     {
-        $id = self::_check_parameter($input, 'id');
-
+        $episode_id = self::_check_parameter($input, 'id');
+        if (!$episode_id) {
+            return;
+        }
         if (AmpConfig::get('podcast') && Access::check('interface', 75)) {
-            $episode = new Podcast_Episode(Subsonic_Xml_Data::_getAmpacheId($id));
+            $episode = new Podcast_Episode(Subsonic_Xml_Data::_getAmpacheId($episode_id));
             if ($episode->id !== null) {
                 $episode->gather();
                 $response = Subsonic_Xml_Data::addSubsonicResponse('downloadpodcastepisode');
@@ -2030,8 +2067,8 @@ class Subsonic_Api
      */
     public static function jukeboxcontrol($input)
     {
-        $action = self::_check_parameter($input, 'action');
-        $id     = $input['id'] ?? array();
+        $action    = self::_check_parameter($input, 'action');
+        $object_id = $input['id'] ?? array();
 
         $response = Subsonic_Xml_Data::addError(Subsonic_Xml_Data::SSERROR_DATA_NOTFOUND, '', 'jukeboxcontrol');
         debug_event(__CLASS__, 'Using Localplay controller: ' . AmpConfig::get('localplay_controller'), 5);
@@ -2066,14 +2103,14 @@ class Subsonic_Api
                 // Intentional break fall-through
                 case 'add':
                     $user = User::get_from_username($input['u']);
-                    if ($id) {
-                        if (!is_array($id)) {
-                            $rid   = array();
-                            $rid[] = $id;
-                            $id    = $rid;
+                    if ($object_id) {
+                        if (!is_array($object_id)) {
+                            $rid       = array();
+                            $rid[]     = $object_id;
+                            $object_id = $rid;
                         }
 
-                        foreach ($id as $song_id) {
+                        foreach ($object_id as $song_id) {
                             $url = null;
 
                             if (Subsonic_Xml_Data::_isSong($song_id)) {
@@ -2439,6 +2476,9 @@ class Subsonic_Api
     {
         $object_id = self::_check_parameter($input, 'id');
         $position  = self::_check_parameter($input, 'position');
+        if (!$object_id || !$position) {
+            return;
+        }
         $comment   = $input['comment'];
         $type      = Subsonic_Xml_Data::_getAmpacheType($object_id);
 
@@ -2473,10 +2513,13 @@ class Subsonic_Api
      */
     public static function deletebookmark($input)
     {
-        $id   = self::_check_parameter($input, 'id');
-        $type = Subsonic_Xml_Data::_getAmpacheType($id);
+        $object_id = self::_check_parameter($input, 'id');
+        if (!$object_id) {
+            return;
+        }
+        $type = Subsonic_Xml_Data::_getAmpacheType($object_id);
 
-        $bookmark = new Bookmark(Subsonic_Xml_Data::_getAmpacheId($id), $type);
+        $bookmark = new Bookmark(Subsonic_Xml_Data::_getAmpacheId($object_id), $type);
         if ($bookmark->id) {
             static::getBookmarkRepository()->delete($bookmark->getId());
             $response = Subsonic_Xml_Data::addSubsonicResponse('deletebookmark');
